LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/getclientrects-003-expected.txt


FAIL Rectangle count assert_equals: expected 3 but got 2
FAIL #0 left assert_equals: expected 150 but got 300
FAIL #0 top assert_equals: expected 60 but got 10
PASS #0 width
FAIL #0 height assert_equals: expected 40 but got 15
PASS #1 left
FAIL #1 top assert_equals: expected 0 but got -50
PASS #1 width
FAIL #1 height assert_equals: expected 100 but got 155
FAIL #2 left undefined is not an object (evaluating 'rects[2].left')
FAIL #2 top undefined is not an object (evaluating 'rects[2].top')
FAIL #2 width undefined is not an object (evaluating 'rects[2].width')
FAIL #2 height undefined is not an object (evaluating 'rects[2].height')
